SQL: Characters db storage type cleanup No. 5

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -2002,7 +2002,7 @@ void ObjectMgr::LoadGameobjectRespawnTimes()
         Field *fields = result->Fetch();
 
         uint32 loguid       = fields[0].GetUInt32();
-        uint64 respawn_time = fields[1].GetUInt64();
+        uint32 respawn_time = fields[1].GetUInt32();
         uint32 instance     = fields[2].GetUInt32();
 
         mGORespawnTimes[MAKE_PAIR64(loguid,instance)] = time_t(respawn_time);
